Renamed chat3 to chat
