--- server/routes/chat/DEagent.ts
@@ -1,212 +1,0 @@
-import { get_samples } from '#src/termdb.sql.js'
-import type { LlmConfig, DbRows } from '#types'
-import { FILTER_TERM_DEFINITIONS, FILTER_DESCRIPTION, validate_filter, num_filter_cutoff } from './filter.ts'
-import { formatTrainingExamples, checkField, generate_group_name, removeLastOccurrence, readJSONFile } from './utils.ts'
-import { route_to_appropriate_llm_provider } from './routeAPIcall.ts'
-import path from 'path'
-import fs from 'fs'
-
-export async function extract_DE_search_terms_from_query(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	ds: any,
-	testing: boolean,
-	aiFilesDir: string
-) {
-	if (ds?.queries?.rnaseqGeneCount) {
-		//const SchemaConfig = {
-		//    path: path.resolve('#types'),
-		//    // Path to your tsconfig (required for proper type resolution)
-		//    tsconfig: path.resolve(serverconfig.binpath, '../tsconfig.json'),
-		//    // Name of the exported type we want to convert
-		//    type: 'DEType',
-		//    // Only expose exported symbols (default)
-		//    expose: 'export' as 'export' | 'all' | 'none' | undefined,
-		//    // Put the whole schema under a top‑level $ref (optional but convenient)
-		//    topRef: true,
-		//    // Turn off type‑checking for speed (set to true if you want full checks)
-		//    skipTypeCheck: true
-		//}
-		//const generator: SchemaGenerator = createGenerator(SchemaConfig)
-		//const Schema = generator.createSchema(SchemaConfig.type) // This commented out code generates the JSON schema below
-
-		const Schema = {
-			$schema: 'http://json-schema.org/draft-07/schema#',
-			$ref: '#/definitions/DEType',
-			definitions: {
-				DEType: {
-					type: 'object',
-					properties: {
-						group1: {
-							type: 'array',
-							items: { $ref: '#/definitions/FilterTerm' },
-							description: 'Name of group1 which is an array of filter terms'
-						},
-						group2: {
-							type: 'array',
-							items: { $ref: '#/definitions/FilterTerm' },
-							description: 'Name of group2 which is an array of filter terms'
-						},
-						method: {
-							type: 'string',
-							enum: ['edgeR', 'limma', 'wilcoxon'],
-							description: 'Method used for carrying out differential gene expression analysis'
-						}
-					},
-					required: ['group1', 'group2'],
-					additionalProperties: false
-				},
-				...FILTER_TERM_DEFINITIONS
-			}
-		} // This JSON schema is generated by ts-json-schema-generator. When DEType is updated, please update this schema by uncommenting the above code and running it locally
-		//mayLog('DEType Schema:', JSON.stringify(Schema))
-
-		// Read DE agent-specific JSON file
-		if (!fs.existsSync(path.join(aiFilesDir, 'differentialAnalysis.json')))
-			throw 'DE agent file is not specified for dataset:' + ds.label
-		const DE_ds = await readJSONFile(path.join(aiFilesDir, 'differentialAnalysis.json'))
-		if (!DE_ds) throw 'DE information is not present in the dataset file.'
-		if (DE_ds.TrainingData.length == 0) throw 'No training data is provided for the DE agent.'
-
-		const training_data = formatTrainingExamples(DE_ds.TrainingData)
-
-		const system_prompt =
-			'I am an assistant that extracts the groups from the user prompt to carry out differential gene expression. The final output must be in the following JSON with NO extra comments. The schema is as follows: ' +
-			JSON.stringify(Schema) +
-			' . "group1" and "group2" fields are compulsory. Both "group1" and "group2" consist of an array of filter variables. ' +
-			FILTER_DESCRIPTION +
-			checkField(dataset_json.DatasetPrompt) +
-			checkField(DE_ds.SystemPrompt) +
-			'The sqlite db in plain language is as follows:\n' +
-			dataset_db_output.rag_docs.join(',') +
-			' training data is as follows:' +
-			training_data +
-			' Question: {' +
-			prompt +
-			'} answer:'
-
-		const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-		if (testing) {
-			// When testing, send raw LLM response
-			const test_response = JSON.parse(response)
-			test_response.plot = 'dge'
-			test_response.type = 'plot'
-			return test_response
-		} else {
-			// In actual production (inside PP) send LLM output for validation
-			return await validate_DE_response(response, ds, dataset_db_output.db_rows)
-		}
-	} else {
-		return { type: 'text', text: 'Differential gene expression not supported for this dataset' }
-	}
-}
-
-async function validate_DE_response(response: string, ds: any, db_rows: DbRows[]) {
-	const response_type = JSON.parse(response)
-	let text = ''
-	let group1: any
-	let samples1lst: any
-	const name1 = generate_group_name(response_type.group1, db_rows)
-	if (!response_type.group1) {
-		text += 'group1 not present in DE output'
-	} else {
-		// Validate filter terms
-		const validated_filters = validate_filter(response_type.group1, ds, name1)
-		if (validated_filters.text.length > 0) {
-			text += validated_filters.text
-		} else {
-			const samples1 = await get_samples({ filter: validated_filters.simplefilter }, ds, true) // true is to bypass permission check
-			samples1lst = samples1.map((item: any) => ({
-				sampleId: item.id,
-				sample: item.name
-			}))
-			group1 = {
-				name: name1,
-				in: true,
-				values: samples1lst
-			}
-		}
-	}
-	let group2: any
-	let samples2lst: any
-	const name2 = generate_group_name(response_type.group2, db_rows)
-	if (!response_type.group2) {
-		text += 'group2 not present in DE output'
-	} else {
-		const validated_filters = validate_filter(response_type.group2, ds, name2)
-		if (validated_filters.text.length > 0) {
-			text += validated_filters.text
-		} else {
-			const samples2 = await get_samples({ filter: validated_filters.simplefilter }, ds, true) // true is to bypass permission check
-			samples2lst = samples2.map((item: any) => ({
-				sampleId: item.id,
-				sample: item.name
-			}))
-			group2 = {
-				name: name2,
-				in: true,
-				values: samples2lst
-			}
-		}
-	}
-
-	// Parse DE method from LLM output
-	let settings: any
-	if (response_type.method) {
-		if (response_type.method == 'edgeR' || response_type.method == 'limma' || response_type.method == 'wilcoxon') {
-			settings = { volcano: { method: response_type.method } }
-		} else {
-			text += 'Unknown DE method: ' + response_type.method
-		}
-	}
-
-	if (text.length > 0) {
-		text = removeLastOccurrence(
-			text,
-			'For now, the maximum number of filter terms supported through the chatbot is ' + num_filter_cutoff // Remove duplicated statements in error message
-		)
-		return { type: 'text', text: text }
-	} else {
-		const pp_plot_json: any = { childType: 'volcano', termType: 'geneExpression', chartType: 'differentialAnalysis' }
-		const groups = [group1, group2]
-		const tw = {
-			q: {
-				groups
-			},
-			term: {
-				name: name1 + ' vs ' + name2,
-				type: 'samplelst',
-				values: {
-					[name1]: {
-						color: 'purple',
-						key: name1,
-						label: name1,
-						list: samples1lst
-					},
-					[name2]: {
-						color: 'blue',
-						key: name2,
-						label: name2,
-						list: samples2lst
-					}
-				}
-			}
-		}
-
-		pp_plot_json.state = {
-			customTerms: [
-				{
-					name: name1 + ' vs ' + name2,
-					tw: tw
-				}
-			],
-			groups: groups
-		}
-		pp_plot_json.samplelst = { groups }
-		pp_plot_json.tw = tw
-		pp_plot_json.settings = settings
-		return { type: 'plot', plot: pp_plot_json }
-	}
-}
