+++ server/routes/chat/genedatatype.ts
