+++ server/routes/chat/scaffold.ts
@@ -16,7 +16,7 @@ import type {
 	Entity
 } from './scaffoldTypes.ts'
 import { extractGenesFromPrompt } from './utils.ts'
-import { classifyGeneDataType } from './genedatatypeagent.ts'
+import { classifyGeneDataType } from './genedatatype.ts'
 import { determineAmbiguousGenePrompt } from './determineAmbiguousGene.ts'
 import { evaluateFilterTerm, phrase2entitytw, collectLeaves, type FilterTreeResult } from './phrase2entity.ts'
 import { getTermObj, type Value } from './entity2termObj.ts'
