+++ server/routes/termdb.chat.ts
@@ -1,189 +1,127 @@
-import fs from 'fs'
-import { ezFetch } from '#shared'
-import { get_samples } from '#src/termdb.sql.js'
-//import { createGenerator } from 'ts-json-schema-generator'
-//import type { SchemaGenerator } from 'ts-json-schema-generator'
-//import path from 'path'
-import type { ChatRequest, ChatResponse, LlmConfig, DbRows, DbValue, ClassificationType, FilterTerm } from '#types'
-import { extractResourceResponse } from './chat/resource.ts'
-import serverconfig from '../src/serverconfig.js'
+import type { ChatRequest, ChatResponse, LlmConfig, QueryClassification } from '#types'
+// import { ambiguousPoints } from '#types'
 import { mayLog } from '#src/helpers.ts'
-import Database from 'better-sqlite3'
 import { formatElapsedTime } from '#shared'
+import { readJSONFile, parse_geneset_db, getChatRelatedPlotTypes } from './chat/utils.ts'
+import { classifyQuery } from './chat/classify1.ts'
+import { classifyPlotType } from './chat/plot.ts'
+import { classifyNotPlot } from './chat/classify2.ts'
+import { inferScaffold } from './chat/scaffold.ts'
+import serverconfig from '../src/serverconfig.js'
+import { getDsAllowedTermTypes } from './termdb.config.ts'
+import { phrase2entity } from './chat/phrase2entity.ts'
+import { inferTermObjFromEntity } from './chat/entity2termObj.ts'
+import { resolveToTwTvs } from './chat/entity2twTvs.ts'
+import { answerDataQueries } from './chat/dataQueries.ts'
+import path from 'path'
+import fs from 'fs'
+import type { Scaffold, Phrase2EntityResult, SummaryScaffold } from './chat/scaffoldTypes.ts'
+import { resolveToPlotState } from './chat/scaffold2state.ts'
 
-const num_filter_cutoff = 3 // The maximum number of filter terms that can be entered and parsed using the chatbot
+/*
+async function doOmnisearch(prompt: string, q, genome: any, ds: any) {
+        const matches = { equals: [], startsWith: [], startsWord: [], includes: [] }
 
-/** Shared JSON Schema definitions for filter terms, used by DE, Summary, and Matrix agents. */
-const FILTER_TERM_DEFINITIONS = {
-	FilterTerm: {
-		anyOf: [{ $ref: '#/definitions/CategoricalFilterTerm' }, { $ref: '#/definitions/NumericFilterTerm' }]
-	},
-	CategoricalFilterTerm: {
-		type: 'object',
-		properties: {
-			term: { type: 'string', description: 'Name of categorical term' },
-			category: { type: 'string', description: 'The category of the term' },
-			join: {
-				type: 'string',
-				enum: ['and', 'or'],
-				description:
-					'join term to be used only when there is more than one filter term and should be placed from the 2nd filter term onwards describing how it connects to the previous term'
-			}
-		},
-		required: ['term', 'category'],
-		additionalProperties: false
-	},
-	NumericFilterTerm: {
-		type: 'object',
-		properties: {
-			term: { type: 'string', description: 'Name of numeric term' },
-			start: { type: 'number', description: 'start position (or lower limit) of numeric term' },
-			stop: { type: 'number', description: 'stop position (or upper limit) of numeric term' },
-			join: {
-				type: 'string',
-				enum: ['and', 'or'],
-				description:
-					'join term to be used only when there is more than one filter term and should be placed from the 2nd filter term onwards describing how it connects to the previous term'
-			}
-		},
-		required: ['term'],
-		additionalProperties: false
-	}
+        // to allow search to work, must unescape special char, e.g. %20 to space
+        // const str = decodeURIComponent(q.findterm).toUpperCase()
+
+        let terms: any = []
+
+        let termdb:any
+        if (ds) {
+                // matches with dataset
+                if (ds?.cohort?.termdb) termdb = ds.cohort.termdb
+                else throw '.cohort.termdb not found on this dataset'
+        }
+        try {
+                console.log(`Running omnisearch for prompt: "${prompt}" on dataset: "${ds.label}"...`)
+                console.log(`q.cohortStr: "${q.cohortStr}", q.usecase: "${q.usecase}", q.treeFilter: ${JSON.stringify(q.treeFilter)}`)
+                const _terms = await termdb.q.findTermByName(prompt, q.cohortStr)
+                console.log(`Omnisearch found ${_terms.length} terms matching the prompt "${prompt}" in dataset "${ds.label}".`)
+                // console.log('Sample of matched terms:', _terms.slice(0, 5))
+                terms.push(..._terms.map(copy_term))
+        } catch (e) {
+                // if (e.stack) console.log(e.stack)
+                throw e 
+        }
+        return terms
 }
-
-/** Format few-shot training examples into a prompt string. */
-function formatTrainingExamples(trainingData: { question: string; answer: any }[]): string {
-	return trainingData
-		.map(
-			(td, i) =>
-				'Example question' +
-				(i + 1).toString() +
-				': ' +
-				td.question +
-				' Example answer' +
-				(i + 1).toString() +
-				':' +
-				JSON.stringify(td.answer)
-		)
-		.join(' ')
-}
-
-/** Shared natural language description of filter term types for LLM prompts. */
-const FILTER_DESCRIPTION =
-	'There are two kinds of filter variables: "Categorical" and "Numeric". ' +
-	'"Categorical" variables are those variables which can have a fixed set of values e.g. gender, race. ' +
-	'They are defined by the "CategoricalFilterTerm" which consists of "term" (a field from the sqlite3 db) and "category" (a value of the field from the sqlite db). ' +
-	'"Numeric" variables are those which can have any numeric value. ' +
-	'They are defined by "NumericFilterTerm" and contain the subfields "term" (a field from the sqlite3 db), ' +
-	'"start" an optional filter which is defined when a lower cutoff is defined in the user input for the numeric variable and ' +
-	'"stop" an optional filter which is defined when a higher cutoff is defined in the user input for the numeric variable. '
-
-/** Extract gene names from user prompt that exist in the gene database. */
-function extractGenesFromPrompt(prompt: string, genes_list: string[]): string[] {
-	const words = prompt
-		.replace(/[^a-zA-Z0-9\s]/g, '')
-		.split(/\s+/)
-		.map(str => str.toLowerCase())
-	return words.filter(item => genes_list.includes(item))
-}
-
-/**
- * Resolve the appropriate chart childType based on the data categories of term and term2,
- * and an optional user-requested override from the LLM output.
- *
- * Returns { childType } on success, or { error } if the user requested an invalid chart type.
- * Also returns { bothNumeric: true } when both terms are numeric so the caller can
- * apply discretization for violin/boxplot.
- */
-
-type TermCategory = 'categorical' | 'float' | 'integer' | undefined
-
-// Default chart type for each (normalized cat1, normalized cat2) combination
-const CHILD_TYPE_DEFAULTS: Record<string, string> = {
-	'categorical:undefined': 'barchart',
-	'numeric:undefined': 'violin',
-	'categorical:categorical': 'barchart',
-	'numeric:categorical': 'violin',
-	'categorical:numeric': 'violin',
-	'numeric:numeric': 'sampleScatter'
-}
-
-// Overrides that are invalid (produce an error) for a given combination
-const CHILD_TYPE_INVALID: Record<string, Set<string>> = {
-	'categorical:undefined': new Set(['violin', 'boxplot', 'sampleScatter']),
-	'categorical:categorical': new Set(['violin', 'boxplot', 'sampleScatter'])
-}
-
-function resolveChildType(
-	cat1: TermCategory,
-	cat2: TermCategory,
-	llmChildType: string | undefined
-): { childType?: string; error?: string; bothNumeric?: boolean } {
-	const norm1 = cat1 == 'float' || cat1 == 'integer' ? 'numeric' : cat1 || 'undefined'
-	const norm2 = cat2 == 'float' || cat2 == 'integer' ? 'numeric' : cat2 || 'undefined'
-	const key = norm1 + ':' + norm2
-	const defaultType = CHILD_TYPE_DEFAULTS[key]
-	if (!defaultType) {
-		// Unknown category combination — should not happen, fall back to barchart
-		return { childType: 'barchart' }
-	}
-	const invalid = CHILD_TYPE_INVALID[key]
-
-	if (llmChildType && invalid && invalid.has(llmChildType)) {
-		return {
-			error:
-				'Invalid plot type supplied by the user: ' +
-				llmChildType +
-				'. For ' +
-				key.replace(':', ' and ') +
-				' variables the plot type should always be ' +
-				defaultType
-		}
-	}
-
-	return {
-		childType: llmChildType || defaultType,
-		bothNumeric: norm1 == 'numeric' && norm2 == 'numeric'
-	}
-}
+*/
 
 export function init({ genomes }) {
 	return async (req, res) => {
 		const q: ChatRequest = req.query
 		try {
-			const g = genomes[q.genome]
-			if (!g) throw 'invalid genome'
-			const ds = g.datasets?.[q.dslabel]
+			const genome = genomes[q.genome]
+			if (!genome) throw 'invalid genome'
+			const ds = genome.datasets?.[q.dslabel]
 			if (!ds) throw 'invalid dslabel'
-			const serverconfig_ds_entries = serverconfig.genomes
-				.find(genome => genome.name == q.genome)
-				.datasets.find(dslabel => dslabel.name == ds.label)
+			// check if ds supports termdb chat
+			if (!ds.queries.chat) {
+				return res.send({
+					type: 'text',
+					text: 'Only search functionality supported for this data. No chat functionality supported.'
+				})
+			}
 
-			if (!serverconfig_ds_entries.aifiles) {
-				throw 'aifiles are missing for chatbot to work'
+			const aiFilesDir = serverconfig.binpath + '/../../dataset/ai/' + q.dslabel // This is the directory where the AI JSON files are stored for this dataset. This will use this as the base directory for resolving all agent file paths specified in the dataset JSON file.
+			let agentFiles: string[] = []
+			try {
+				// Read dataset JSON file
+				agentFiles = await fs.readdirSync(aiFilesDir).filter(file => file.endsWith('.json'))
+			} catch (err: any) {
+				if (err.code === 'ENOENT') throw new Error(`Directory not found: ${aiFilesDir}`)
+				if (err.code === 'ENOTDIR') throw new Error(`Path is not a directory: ${aiFilesDir}`)
+				throw err
 			}
 
 			const llm = serverconfig.llm
 			if (!llm) throw 'serverconfig.llm is not configured'
-			if (llm.provider !== 'SJ' && llm.provider !== 'ollama') {
-				throw "llm.provider must be 'SJ' or 'ollama'"
+			if (
+				llm.provider !== 'SJ' &&
+				llm.provider !== 'ollama' &&
+				llm.provider !== 'huggingface' &&
+				(llm.provider as string) !== 'azure'
+			) {
+				throw "llm.provider must be 'SJ', 'ollama', 'huggingface', or 'azure'"
 			}
 
-			const dataset_db = serverconfig.tpmasterdir + '/' + ds.cohort.db.file
-			const genedb = serverconfig.tpmasterdir + '/' + g.genedb.dbfile
-			// Read dataset JSON file
-			const dataset_json: any = await readJSONFile(serverconfig_ds_entries.aifiles)
-			const testing = false // This toggles validation of LLM output. In this script, this will ALWAYS be false since we always want validation of LLM output, only for testing we this variable to true
+			// This toggles validation of LLM output. In this script, this will ALWAYS be false since we always want validation of LLM output,
+			// only for testing we set this variable to true
+			// const testing = false
+
+			// Access the cohort key/label
+			let rawFilter: any
+			if (typeof q.filter === 'string') {
+				try {
+					rawFilter = JSON.parse(q.filter)
+				} catch (e) {
+					throw new Error('Failed to parse filter JSON string: ' + e)
+				}
+			} else {
+				rawFilter = q.filter
+			}
+			const filter: any = rawFilter && typeof rawFilter === 'object' ? rawFilter : {}
+			const lst = Array.isArray(filter.lst) ? filter.lst : []
+			const cohortFilter = lst.find((item: any) => item.tag === 'cohortFilter')
+			const cohortKey = cohortFilter ? cohortFilter.tvs.values[0].key : ''
+			const supportedPlotTypes = ds.cohort.termdb.q?.getSupportedChartTypes(req)?.[cohortKey]
+			const chatSupportedPlotTypes = getChatRelatedPlotTypes(supportedPlotTypes)
+			const genedb = serverconfig.tpmasterdir + '/' + genome.genedb.dbfile
+			const allowedTermTypes = getDsAllowedTermTypes(ds) as string[]
 			const ai_output_json = await run_chat_pipeline(
 				q.prompt,
 				llm,
-				serverconfig.aiRoute,
-				dataset_json,
-				testing,
-				dataset_db,
+				ds,
 				genedb,
-				ds
+				agentFiles,
+				aiFilesDir,
+				chatSupportedPlotTypes,
+				allowedTermTypes,
+				genome
+				// 	testing
 			)
+			mayLog('From init: Final AI output JSON:', JSON.stringify(ai_output_json))
 			res.send(ai_output_json as ChatResponse)
 		} catch (e: any) {
 			if (e.stack) mayLog(e.stack)
@@ -193,1269 +131,156 @@ export function init({ genomes }) {
 }
 
 export async function run_chat_pipeline(
-	user_prompt: string,
+	userPrompt: string,
 	llm: LlmConfig,
-	aiRoute: string,
-	dataset_json: any,
-	testing: boolean,
-	dataset_db: string,
+	ds: any,
 	genedb: string,
-	ds: any
+	agentFiles: string[],
+	aiFilesDir: string,
+	supportedPlotTypes: string[],
+	allowedTermTypes: string[],
+	genome: any
+	// testing: boolean
 ) {
+	// Read main.json file
+	if (!fs.existsSync(path.join(aiFilesDir, 'main.json')))
+		throw 'Main data file is not specified for dataset:' + ds.label
+	const dataset_json: any = await readJSONFile(path.join(aiFilesDir, 'main.json'))
+
+	// Plot vs Not-plot classification
 	const time1 = new Date().valueOf()
-	const class_response: ClassificationType = await classify_query_by_dataset_type(
-		user_prompt,
-		llm,
-		aiRoute,
-		dataset_json
-	)
-	let ai_output_json: any
+	const class_response: QueryClassification = await classifyQuery(userPrompt, llm)
 	mayLog('Time taken for classification:', formatElapsedTime(Date.now() - time1))
-	if (class_response.type == 'none') {
-		ai_output_json = {
-			type: 'text',
-			text: 'Your query does not appear to be related to the available data visualizations. Please try rephrasing your question.'
-		}
-	} else if (class_response.type == 'resource') {
-		const time1 = new Date().valueOf()
-		ai_output_json = await extractResourceResponse(user_prompt, llm, dataset_json)
-		mayLog('Time taken for resource agent:', formatElapsedTime(Date.now() - time1))
-	} else if (class_response.type == 'plot') {
-		const classResult = class_response.plot
-		mayLog('classResult:', classResult)
-		// Parse DBs once and pass to whichever agent runs
-		const dataset_db_output = await parse_dataset_db(dataset_db)
-		const genes_list = dataset_json.hasGeneExpression ? await parse_geneset_db(genedb) : []
-		if (classResult == 'summary') {
-			const time1 = new Date().valueOf()
-			ai_output_json = await extract_summary_terms(
-				user_prompt,
-				llm,
-				dataset_db_output,
-				dataset_json,
-				genes_list,
-				ds,
-				testing
-			)
-			mayLog('Time taken for summary agent:', formatElapsedTime(Date.now() - time1))
-		} else if (classResult == 'dge') {
-			const time1 = new Date().valueOf()
-			ai_output_json = await extract_DE_search_terms_from_query(
-				user_prompt,
-				llm,
-				dataset_db_output,
-				dataset_json,
-				ds,
-				testing
-			)
-			mayLog('Time taken for DE agent:', formatElapsedTime(Date.now() - time1))
-		} else if (classResult == 'survival') {
-			ai_output_json = { type: 'text', text: 'survival agent has not been implemented yet' }
-		} else if (classResult == 'matrix') {
-			const time1 = new Date().valueOf()
-			ai_output_json = await extract_matrix_search_terms_from_query(
-				user_prompt,
-				llm,
-				dataset_db_output,
-				dataset_json,
-				genes_list,
-				ds,
-				testing
-			)
-			mayLog('Time taken for matrix agent:', formatElapsedTime(Date.now() - time1))
-		} else if (classResult == 'sampleScatter') {
-			const time1 = new Date().valueOf()
-			ai_output_json = await extract_samplescatter_terms_from_query(
-				user_prompt,
-				llm,
-				dataset_db_output,
-				dataset_json,
-				genes_list,
-				ds,
-				testing
-			)
-			mayLog('Time taken for sampleScatter agent:', formatElapsedTime(Date.now() - time1))
+
+	let ai_output_json: any
+	if (class_response.type === 'notplot') {
+		// If Not-plot: Resource or None classification
+		const time2 = new Date().valueOf()
+		const notPlotResult = await classifyNotPlot(userPrompt, llm, agentFiles, aiFilesDir) //, allowedTermTypes)
+		mayLog('Time taken for classify2:', formatElapsedTime(Date.now() - time2))
+
+		if (notPlotResult.type === 'html') {
+			ai_output_json = notPlotResult
 		} else {
-			// Will define all other agents later as desired
-			ai_output_json = { type: 'text', text: 'Unknown classification value' }
+			ai_output_json = {
+				type: 'text',
+				text: 'Your query does not appear to be related to the available data visualizations. Please try rephrasing your question.'
+			}
 		}
-	} else {
-		// Should not happen
-		ai_output_json = { type: 'text', text: 'Unknown classification type' }
+	} else if (class_response.type === 'binaryQuery') {
+		const answer = await answerDataQueries(userPrompt, llm, allowedTermTypes)
+		if (!answer) throw "Couldn't decide if this is data related query!"
+		mayLog('Data Binary Query: ', answer)
+		ai_output_json = answer
+	} else if (class_response.type === 'plot') {
+		let time = new Date().valueOf()
+		const plotType = await classifyPlotType(userPrompt, llm)
+		mayLog('Time taken to classify plot type:', formatElapsedTime(Date.now() - time))
+
+		// Check if the classified plot type is supported by this dataset
+		if (!supportedPlotTypes.includes(plotType)) {
+			const log = 'Plot type: "' + plotType + '" is not supported.'
+			ai_output_json = {
+				type: 'text',
+				text: log
+			}
+			mayLog(log)
+			return ai_output_json
+		}
+
+		const genes_list = await parse_geneset_db(genedb)
+
+		// If supported plot type, figure out the scaffold according to the plot type
+		mayLog('#################################################')
+		mayLog('####### First phase: Infer Plot Scaffolds #######')
+		mayLog('#################################################')
+		time = new Date().valueOf()
+		const dataset_db = serverconfig.tpmasterdir + '/' + ds.cohort.db.file
+		const scaffoldResult = await inferScaffold(
+			userPrompt,
+			plotType,
+			llm,
+			genome,
+			genes_list,
+			allowedTermTypes,
+			dataset_json,
+			ds,
+			dataset_db
+		)
+		mayLog('ScaffoldResult: ', scaffoldResult)
+		if (
+			(plotType === 'hiercluster' && 'plot' in scaffoldResult && scaffoldResult.type === 'plot') ||
+			('text' in scaffoldResult && scaffoldResult.type === 'text')
+		) {
+			// In case of geneExpression clustering, the plot state is generated through a single LLM call and invoking downstream steps is not necessary.
+			return scaffoldResult
+		}
+		mayLog('Time taken to infer scaffold:', formatElapsedTime(Date.now() - time))
+
+		if (!scaffoldResult)
+			throw 'Scaffold result is empty or undefined, which is unexpected. Please check the inferScaffold agent for potential issues.'
+		if ('type' in scaffoldResult && scaffoldResult.type === 'text') {
+			return scaffoldResult // Return msg/error
+		}
+
+		const subplotType =
+			(scaffoldResult as Scaffold).plotType === 'summary' ? (scaffoldResult as SummaryScaffold).chartType : undefined
+
+		mayLog('#################################################')
+		mayLog("####### Second phase: From Scaffolds's phrases infer Entities #######")
+		mayLog('#################################################')
+		time = new Date().valueOf()
+		const phrase2entityResult = await phrase2entity(
+			scaffoldResult as Scaffold,
+			plotType,
+			llm,
+			genes_list,
+			dataset_json,
+			ds,
+			genome
+		)
+		mayLog('Time taken to phrase 2 entity:', formatElapsedTime(Date.now() - time))
+		if ('type' in phrase2entityResult && phrase2entityResult.type === 'text') {
+			return phrase2entityResult // Return msg/error
+		}
+		mayLog(phrase2entityResult)
+
+		mayLog('#################################################')
+		mayLog('####### Third phase: From Entities infer Term Objects #######')
+		mayLog('#################################################')
+		time = new Date().valueOf()
+		const termObj = await inferTermObjFromEntity(
+			phrase2entityResult as Phrase2EntityResult,
+			plotType,
+			llm,
+			dataset_db,
+			genes_list,
+			genome
+		)
+		mayLog('Time taken to infer term objects:', formatElapsedTime(Date.now() - time))
+		mayLog('Inferred termObj from entity:', JSON.stringify(termObj))
+
+		mayLog('#################################################')
+		mayLog('####### Fourth phase: From Term Objects to TwTvs Objects #######')
+		mayLog('#################################################')
+		time = new Date().valueOf()
+		const twTvsObj = await resolveToTwTvs(termObj, plotType, llm, dataset_db, genome)
+		mayLog('Time taken to resolve to TwTvs object from termObj:', formatElapsedTime(Date.now() - time))
+		if ('type' in twTvsObj && twTvsObj.type === 'text') {
+			return twTvsObj // Return msg/error
+		}
+		mayLog('twTvsObj:', twTvsObj)
+
+		mayLog('#################################################')
+		mayLog('####### Fifth/Final phase: From TwTvs Objects to Plot States #######')
+		mayLog('#################################################')
+		time = new Date().valueOf()
+		ai_output_json = resolveToPlotState(twTvsObj, plotType, subplotType)
+		mayLog('Time taken to resolve to plot state:', formatElapsedTime(Date.now() - time))
+		// TODO: might need a validation step here to check if the scaffoldResult contains valid term types that
+		// are present in the dataset and compatible with the plot type, and if not return an error message to the user.
+		// This is a bit complex because it requires cross-referencing the inferred scaffold with the dataset's schema and allowed term types,
+		//  but it would improve robustness.
 	}
 	return ai_output_json
-}
-
-async function call_ollama(prompt: string, model_name: string, apilink: string) {
-	const temperature = 0.01
-	const top_p = 0.95
-	const timeout = 200000
-	const payload = {
-		model: model_name,
-		messages: [{ role: 'user', content: prompt }],
-		raw: false,
-		stream: false,
-		keep_alive: 15, //Keep the LLM loaded for 15mins
-		options: {
-			top_p: top_p,
-			temperature: temperature,
-			num_ctx: 10000
-		}
-	}
-
-	try {
-		const result = await ezFetch(apilink + '/api/chat', {
-			method: 'POST',
-			body: payload, // ezfetch automatically stringifies objects
-			headers: { 'Content-Type': 'application/json' },
-			timeout: { request: timeout } // ezfetch accepts milliseconds directly
-		})
-		if (result && result.message && result.message.content && result.message.content.length > 0)
-			return result.message.content
-		else {
-			throw 'Error: Received an unexpected response format:' + result
-		}
-	} catch (error) {
-		throw 'Ollama API request failed:' + error
-	}
-}
-
-async function call_sj_llm(prompt: string, model_name: string, apilink: string) {
-	const temperature = 0.01
-	const top_p = 0.95
-	const timeout = 200000
-	const max_new_tokens = 512
-	const payload = {
-		inputs: [
-			{
-				model_name: model_name,
-				inputs: {
-					text: prompt,
-					max_new_tokens: max_new_tokens,
-					temperature: temperature,
-					top_p: top_p
-				}
-			}
-		]
-	}
-
-	try {
-		const response = await ezFetch(apilink, {
-			method: 'POST',
-			body: payload, // ezfetch automatically stringifies objects
-			headers: { 'Content-Type': 'application/json' },
-			timeout: { request: timeout } // ezfetch accepts milliseconds directly
-		})
-		if (response.outputs && response.outputs[0] && response.outputs[0].generated_text) {
-			const result = response.outputs[0].generated_text
-			return result
-		} else {
-			throw 'Error: Received an unexpected response format:' + response
-		}
-	} catch (error) {
-		throw 'SJ API request failed:' + error
-	}
-}
-
-async function route_to_appropriate_llm_provider(template: string, llm: LlmConfig): Promise<string> {
-	let response: string
-	if (llm.provider == 'SJ') {
-		// Local SJ server
-		response = await call_sj_llm(template, llm.modelName, llm.api)
-	} else if (llm.provider == 'ollama') {
-		// Ollama server
-		response = await call_ollama(template, llm.modelName, llm.api)
-	} else {
-		// Will later add support for azure server also
-		throw 'Unknown LLM provider'
-	}
-	return response
-}
-
-function checkField(sentence: string) {
-	if (!sentence) return ''
-	else return sentence
-}
-
-export async function readJSONFile(file: string) {
-	const json_file = await fs.promises.readFile(file)
-	return JSON.parse(json_file.toString())
-}
-
-async function classify_query_by_dataset_type(user_prompt: string, llm: LlmConfig, aiRoute: string, dataset_json: any) {
-	const data = await readJSONFile(aiRoute)
-	let contents = data['general'] // The general description should be right at the top of the system prompt
-	for (const key of Object.keys(data)) {
-		// Add descriptions of all other agents after the general description
-		if (key != 'general') {
-			contents += data[key]
-		}
-	}
-
-	// Parse out training data from the dataset JSON and add it to a string
-	const classification_ds = dataset_json.charts.find((chart: any) => chart.type == 'Classification')
-	if (!classification_ds) throw 'Classification information is not present in the dataset file.'
-	if (classification_ds.TrainingData.length == 0) throw 'No training data is provided for the classification agent.'
-	let training_data = ''
-	if (classification_ds && classification_ds.TrainingData.length > 0) {
-		contents += checkField(dataset_json.DatasetPrompt) + checkField(classification_ds.SystemPrompt)
-		training_data = formatTrainingExamples(classification_ds.TrainingData)
-	}
-
-	//const SchemaConfig = {
-	//    path: path.resolve('#types'),
-	//    // Path to your tsconfig (required for proper type resolution)
-	//    tsconfig: path.resolve(serverconfig.binpath, '../tsconfig.json'),
-	//    // Name of the exported type we want to convert
-	//    type: 'ClassificationType',
-	//    // Only expose exported symbols (default)
-	//    expose: 'export' as 'export' | 'all' | 'none' | undefined,
-	//    // Put the whole schema under a top‑level $ref (optional but convenient)
-	//    topRef: true,
-	//    // Turn off type‑checking for speed (set to true if you want full checks)
-	//    skipTypeCheck: true
-	//}
-	//const generator: SchemaGenerator = createGenerator(SchemaConfig)
-	//const Schema = JSON.stringify(generator.createSchema(SchemaConfig.type)) // This will be generated at server startup later
-	//mayLog("ClassificationType StringifiedSchema:", Schema)
-
-	const template =
-		contents + ' training data is as follows:' + training_data + ' Question: {' + user_prompt + '} Answer: {answer}'
-
-	const response: string = await route_to_appropriate_llm_provider(template, llm)
-	//if (testing) {
-	//    return { action: 'html', response: response } // In case of classification agent a response could be an html pointing to some weblink. So JSON.parse() will fail
-	//} else {
-	//    return JSON.parse(response)
-	//}
-	return JSON.parse(response)
-}
-
-async function extract_DE_search_terms_from_query(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	ds: any,
-	testing: boolean
-) {
-	if (dataset_json.hasDE) {
-		//const SchemaConfig = {
-		//    path: path.resolve('#types'),
-		//    // Path to your tsconfig (required for proper type resolution)
-		//    tsconfig: path.resolve(serverconfig.binpath, '../tsconfig.json'),
-		//    // Name of the exported type we want to convert
-		//    type: 'DEType',
-		//    // Only expose exported symbols (default)
-		//    expose: 'export' as 'export' | 'all' | 'none' | undefined,
-		//    // Put the whole schema under a top‑level $ref (optional but convenient)
-		//    topRef: true,
-		//    // Turn off type‑checking for speed (set to true if you want full checks)
-		//    skipTypeCheck: true
-		//}
-		//const generator: SchemaGenerator = createGenerator(SchemaConfig)
-		//const Schema = generator.createSchema(SchemaConfig.type) // This commented out code generates the JSON schema below
-		const Schema = {
-			$schema: 'http://json-schema.org/draft-07/schema#',
-			$ref: '#/definitions/DEType',
-			definitions: {
-				DEType: {
-					type: 'object',
-					properties: {
-						group1: {
-							type: 'array',
-							items: { $ref: '#/definitions/FilterTerm' },
-							description: 'Name of group1 which is an array of filter terms'
-						},
-						group2: {
-							type: 'array',
-							items: { $ref: '#/definitions/FilterTerm' },
-							description: 'Name of group2 which is an array of filter terms'
-						},
-						method: {
-							type: 'string',
-							enum: ['edgeR', 'limma', 'wilcoxon'],
-							description: 'Method used for carrying out differential gene expression analysis'
-						}
-					},
-					required: ['group1', 'group2'],
-					additionalProperties: false
-				},
-				...FILTER_TERM_DEFINITIONS
-			}
-		} // This JSON schema is generated by ts-json-schema-generator. When DEType is updated, please update this schema by uncommenting the above code and running it locally
-		//mayLog('DEType Schema:', JSON.stringify(Schema))
-
-		// Parse out training data from the dataset JSON and add it to a string
-		const DE_ds = dataset_json.charts.find((chart: any) => chart.type == 'DE')
-		if (!DE_ds) throw 'DE information is not present in the dataset file.'
-		if (DE_ds.TrainingData.length == 0) throw 'No training data is provided for the DE agent.'
-
-		const training_data = formatTrainingExamples(DE_ds.TrainingData)
-
-		const system_prompt =
-			'I am an assistant that extracts the groups from the user prompt to carry out differential gene expression. The final output must be in the following JSON with NO extra comments. The schema is as follows: ' +
-			JSON.stringify(Schema) +
-			' . "group1" and "group2" fields are compulsory. Both "group1" and "group2" consist of an array of filter variables. ' +
-			FILTER_DESCRIPTION +
-			checkField(dataset_json.DatasetPrompt) +
-			checkField(DE_ds.SystemPrompt) +
-			'The sqlite db in plain language is as follows:\n' +
-			dataset_db_output.rag_docs.join(',') +
-			' training data is as follows:' +
-			training_data +
-			' Question: {' +
-			prompt +
-			'} answer:'
-
-		const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-		if (testing) {
-			// When testing, send raw LLM response
-			return { action: 'dge', response: JSON.parse(response) }
-		} else {
-			// In actual production (inside PP) send LLM output for validation
-			return await validate_DE_response(response, ds, dataset_db_output.db_rows)
-		}
-	} else {
-		return { type: 'html', html: 'Differential gene expression not supported for this dataset' }
-	}
-}
-
-async function validate_DE_response(response: string, ds: any, db_rows: DbRows[]) {
-	const response_type = JSON.parse(response)
-	let html = ''
-	let group1: any
-	let samples1lst: any
-	const name1 = generate_group_name(response_type.group1, db_rows)
-	if (!response_type.group1) {
-		html += 'group1 not present in DE output'
-	} else {
-		// Validate filter terms
-		const validated_filters = validate_filter(response_type.group1, ds, name1)
-		if (validated_filters.html.length > 0) {
-			html += validated_filters.html
-		} else {
-			const samples1 = await get_samples({ filter: validated_filters.simplefilter }, ds, true) // true is to bypass permission check
-			samples1lst = samples1.map((item: any) => ({
-				sampleId: item.id,
-				sample: item.name
-			}))
-			group1 = {
-				name: name1,
-				in: true,
-				values: samples1lst
-			}
-		}
-	}
-	let group2: any
-	let samples2lst: any
-	const name2 = generate_group_name(response_type.group2, db_rows)
-	if (!response_type.group2) {
-		html += 'group2 not present in DE output'
-	} else {
-		const validated_filters = validate_filter(response_type.group2, ds, name2)
-		if (validated_filters.html.length > 0) {
-			html += validated_filters.html
-		} else {
-			const samples2 = await get_samples({ filter: validated_filters.simplefilter }, ds, true) // true is to bypass permission check
-			samples2lst = samples2.map((item: any) => ({
-				sampleId: item.id,
-				sample: item.name
-			}))
-			group2 = {
-				name: name2,
-				in: true,
-				values: samples2lst
-			}
-		}
-	}
-
-	// Parse DE method from LLM output
-	let settings: any
-	if (response_type.method) {
-		if (response_type.method == 'edgeR' || response_type.method == 'limma' || response_type.method == 'wilcoxon') {
-			settings = { volcano: { method: response_type.method } }
-		} else {
-			html += 'Unknown DE method: ' + response_type.method
-		}
-	}
-
-	if (html.length > 0) {
-		html = removeLastOccurrence(
-			html,
-			'For now, the maximum number of filter terms supported through the chatbot is ' + num_filter_cutoff // Remove duplicated statements in error message
-		)
-		return { type: 'html', html: html }
-	} else {
-		const pp_plot_json: any = { childType: 'volcano', termType: 'geneExpression', chartType: 'differentialAnalysis' }
-		const groups = [group1, group2]
-		const tw = {
-			q: {
-				groups
-			},
-			term: {
-				name: name1 + ' vs ' + name2,
-				type: 'samplelst',
-				values: {
-					[name1]: {
-						color: 'purple',
-						key: name1,
-						label: name1,
-						list: samples1lst
-					},
-					[name2]: {
-						color: 'blue',
-						key: name2,
-						label: name2,
-						list: samples2lst
-					}
-				}
-			}
-		}
-		pp_plot_json.state = {
-			customTerms: [
-				{
-					name: name1 + ' vs ' + name2,
-					tw: tw
-				}
-			],
-			groups: groups
-		}
-		pp_plot_json.samplelst = { groups }
-		pp_plot_json.tw = tw
-		pp_plot_json.settings = settings
-		return { type: 'plot', plot: pp_plot_json }
-	}
-}
-
-function generate_group_name(filters: any[], db_rows: DbRows[]): string {
-	let name = ''
-	let iter = 0
-	for (const filter of filters) {
-		if (iter > 0 && !filter.join) {
-			// Sometimes the LLM misses join terms. In such cases, hardcoding & operator
-			name += '&'
-		}
-		if (filter.join && filter.join == 'and') {
-			name += '&'
-		}
-		if (filter.join && filter.join == 'or') {
-			name += '|'
-		}
-		if (filter.category) {
-			// Categorical variable
-			name += find_label(filter, db_rows)
-		}
-		if (filter.start) {
-			// Integer or float variable
-			name += filter.term + '>=' + filter.start.toString()
-		}
-		if (filter.stop) {
-			// Integer or float variable
-			name += filter.term + '<=' + filter.stop.toString()
-		}
-		iter += 1
-	}
-	return name
-}
-
-function find_label(filter: any, db_rows: DbRows[]): string {
-	let label = ''
-	for (const row of db_rows) {
-		if (row.name == filter.term) {
-			for (const value of row.values) {
-				if (value.value && value.value.label && filter.category == value.key) {
-					label = value.value.label
-					break
-				}
-			}
-			break
-		}
-	}
-	return label
-}
-
-async function extract_summary_terms(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	genes_list: string[],
-	ds: any,
-	testing: boolean
-) {
-	//const SchemaConfig = {
-	//    path: path.resolve('#types'),
-	//    // Path to your tsconfig (required for proper type resolution)
-	//    tsconfig: path.resolve(serverconfig.binpath, '../tsconfig.json'),
-	//    // Name of the exported type we want to convert
-	//    type: 'SummaryType',
-	//    // Only expose exported symbols (default)
-	//    expose: 'export' as 'export' | 'all' | 'none' | undefined,
-	//    // Put the whole schema under a top‑level $ref (optional but convenient)
-	//    topRef: true,
-	//    // Turn off type‑checking for speed (set to true if you want full checks)
-	//    skipTypeCheck: true
-	//}
-	//const generator: SchemaGenerator = createGenerator(SchemaConfig)
-	//const Schema = JSON.stringify(generator.createSchema(SchemaConfig.type)) // This will be generated at server startup later
-	const Schema = {
-		$schema: 'http://json-schema.org/draft-07/schema#',
-		$ref: '#/definitions/SummaryType',
-		definitions: {
-			SummaryType: {
-				type: 'object',
-				properties: {
-					term: { type: 'string', description: 'Name of 1st term' },
-					term2: { type: 'string', description: 'Name of 2nd term' },
-					simpleFilter: {
-						type: 'array',
-						items: { $ref: '#/definitions/FilterTerm' },
-						description: 'Optional simple filter terms'
-					},
-					childType: {
-						type: 'string',
-						enum: ['violin', 'boxplot', 'sampleScatter', 'barchart'],
-						description:
-							'Optional explicit child type requested by the user. If omitted, the logic of the data types picks the child type.'
-					}
-				},
-				required: ['term', 'simpleFilter'],
-				additionalProperties: false
-			},
-			...FILTER_TERM_DEFINITIONS
-		}
-	} // This JSON schema is generated by ts-json-schema-generator. When SummaryType is updated, please update this schema by uncommenting the above code and running it locally
-
-	const common_genes = extractGenesFromPrompt(prompt, genes_list)
-
-	// Parse out training data from the dataset JSON and add it to a string
-	const summary_ds = dataset_json.charts.find((chart: any) => chart.type == 'Summary')
-	if (!summary_ds) throw 'Summary information is not present in the dataset file.'
-	if (summary_ds.TrainingData.length == 0) throw 'No training data is provided for the summary agent.'
-
-	const training_data = formatTrainingExamples(summary_ds.TrainingData)
-
-	let system_prompt =
-		'I am an assistant that extracts the summary terms from user query. The final output must be in the following JSON format with NO extra comments. The JSON schema is as follows: ' +
-		JSON.stringify(Schema) +
-		' term and term2 (if present) should ONLY contain names of the fields from the sqlite db. The "simpleFilter" field is optional and should contain an array of JSON terms with which the dataset will be filtered. ' +
-		FILTER_DESCRIPTION +
-		checkField(dataset_json.DatasetPrompt) +
-		checkField(summary_ds.SystemPrompt) +
-		'\n The DB content is as follows: ' +
-		dataset_db_output.rag_docs.join(',') +
-		' training data is as follows:' +
-		training_data
-
-	if (dataset_json.hasGeneExpression) {
-		// If dataset has geneExpression data
-		if (common_genes.length > 0) {
-			system_prompt += '\n List of relevant genes are as follows (separated by comma(,)):' + common_genes.join(',')
-		}
-	}
-
-	system_prompt += ' Question: {' + prompt + '} answer:'
-
-	const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-	if (testing) {
-		// When testing, send raw LLM response
-		return { action: 'summary', response: JSON.parse(response) }
-	} else {
-		// In actual production (inside PP) send LLM output for validation
-		return validate_summary_response(response, common_genes, dataset_json, ds)
-	}
-}
-
-function validate_summary_response(response: string, common_genes: string[], dataset_json: any, ds: any) {
-	const response_type = JSON.parse(response)
-	const pp_plot_json: any = { chartType: 'summary' }
-	let html = ''
-	if (response_type.html) html = response_type.html
-	if (!response_type.term) {
-		html += 'term type is not present in summary output'
-		return { type: 'html', html: html }
-	}
-	const term1_validation = validate_term(response_type.term, common_genes, dataset_json, ds)
-	if (term1_validation.html.length > 0) {
-		html += term1_validation.html
-		return { type: 'html', html: html }
-	} else {
-		pp_plot_json.term = term1_validation.term_type
-		if (term1_validation.category == 'float' || term1_validation.category == 'integer') {
-			pp_plot_json.term.q = { mode: 'continuous' }
-		}
-		pp_plot_json.category = term1_validation.category
-	}
-
-	if (response_type.term2) {
-		const term2_validation = validate_term(response_type.term2, common_genes, dataset_json, ds)
-		if (term2_validation.html.length > 0) {
-			html += term2_validation.html
-			return { type: 'html', html: html }
-		} else {
-			pp_plot_json.term2 = term2_validation.term_type
-			if (term2_validation.category == 'float' || term2_validation.category == 'integer') {
-				pp_plot_json.term2.q = { mode: 'continuous' }
-			}
-			pp_plot_json.category2 = term2_validation.category
-		}
-	}
-	/** Based on data types of term and term2, decide the most appropriate chart type.
-	 *  The user can override the default by explicitly mentioning a chart type in their prompt,
-	 *  which the LLM parses into the "childType" field. Invalid overrides produce an error. */
-	const llmChildType =
-		response_type.childType && ['violin', 'boxplot', 'sampleScatter', 'barchart'].includes(response_type.childType)
-			? response_type.childType
-			: undefined
-	const resolved = resolveChildType(pp_plot_json.category, pp_plot_json.category2, llmChildType)
-
-	if (resolved.error) {
-		html += resolved.error
-		return { type: 'html', html: html }
-	} else {
-		pp_plot_json.childType = resolved.childType
-		// For two numeric variables displayed as violin/boxplot, discretize term2
-		if (pp_plot_json.childType == 'barchart') {
-			pp_plot_json.term.q = { mode: 'discrete' }
-			if (pp_plot_json.term2) {
-				pp_plot_json.term2.q = { mode: 'discrete' }
-			}
-		}
-		if (resolved.bothNumeric && (resolved.childType == 'violin' || resolved.childType == 'boxplot')) {
-			pp_plot_json.term2.q = { mode: 'discrete' }
-		}
-	}
-
-	delete pp_plot_json.category
-	if (pp_plot_json.category2) delete pp_plot_json.category2
-
-	if (response_type.simpleFilter && response_type.simpleFilter.length > 0) {
-		const validated_filters = validate_filter(response_type.simpleFilter, ds, '')
-		if (validated_filters.html.length > 0) {
-			html += validated_filters.html
-			return { type: 'html', html: html }
-		} else {
-			pp_plot_json.filter = validated_filters.simplefilter
-		}
-	}
-	return { type: 'plot', plot: pp_plot_json }
-}
-
-async function extract_matrix_search_terms_from_query(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	genes_list: string[],
-	ds: any,
-	testing: boolean
-) {
-	const Schema = {
-		$schema: 'http://json-schema.org/draft-07/schema#',
-		$ref: '#/definitions/MatrixType',
-		definitions: {
-			MatrixType: {
-				type: 'object',
-				properties: {
-					terms: {
-						type: 'array',
-						items: { type: 'string' },
-						description: 'Names of dictionary/clinical terms to include as rows in the matrix'
-					},
-					geneNames: {
-						type: 'array',
-						items: { type: 'string' },
-						description: 'Names of genes to include as gene variant rows in the matrix'
-					},
-					simpleFilter: {
-						type: 'array',
-						items: { $ref: '#/definitions/FilterTerm' },
-						description: 'Optional simple filter terms to restrict the sample set'
-					}
-				},
-				additionalProperties: false
-			},
-			...FILTER_TERM_DEFINITIONS
-		}
-	}
-
-	const common_genes = extractGenesFromPrompt(prompt, genes_list)
-
-	// Parse out training data from the dataset JSON
-	const matrix_ds = dataset_json.charts.filter((chart: any) => chart.type == 'Matrix')
-	if (matrix_ds.length == 0) throw 'Matrix information is not present in the dataset file.'
-	if (matrix_ds[0].TrainingData.length == 0) throw 'No training data is provided for the matrix agent.'
-
-	const training_data = formatTrainingExamples(matrix_ds[0].TrainingData)
-
-	let system_prompt =
-		'I am an assistant that extracts terms and gene names from the user query to create a matrix plot. A matrix plot displays multiple genes and/or clinical variables across samples in a grid layout. The final output must be in the following JSON format with NO extra comments. The JSON schema is as follows: ' +
-		JSON.stringify(Schema) +
-		' The "terms" field should ONLY contain names of clinical/dictionary fields from the sqlite db. The "geneNames" field should ONLY contain gene names. At least one of "terms" or "geneNames" must be provided. The "simpleFilter" field is optional and should contain an array of JSON terms with which the dataset will be filtered. ' +
-		FILTER_DESCRIPTION +
-		checkField(dataset_json.DatasetPrompt) +
-		checkField(matrix_ds[0].SystemPrompt) +
-		'\n The DB content is as follows: ' +
-		dataset_db_output.rag_docs.join(',') +
-		' training data is as follows:' +
-		training_data
-
-	if (dataset_json.hasGeneExpression && common_genes.length > 0) {
-		system_prompt += '\n List of relevant genes are as follows (separated by comma(,)):' + common_genes.join(',')
-	}
-
-	system_prompt += ' Question: {' + prompt + '} answer:'
-
-	const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-	if (testing) {
-		return { action: 'matrix', response: JSON.parse(response) }
-	} else {
-		return validate_matrix_response(response, common_genes, dataset_json, ds)
-	}
-}
-
-function validate_matrix_response(response: string, common_genes: string[], dataset_json: any, ds: any) {
-	const response_type = JSON.parse(response)
-	const pp_plot_json: any = { chartType: 'matrix' }
-	let html = ''
-
-	if (response_type.html) html = response_type.html
-
-	// Must have at least one of terms or geneNames
-	if (
-		(!response_type.terms || response_type.terms.length == 0) &&
-		(!response_type.geneNames || response_type.geneNames.length == 0)
-	) {
-		html += 'At least one clinical term or gene name is required for a matrix plot'
-	}
-
-	// Validate dictionary terms — use shorthand { id } at tw top level
-	const twLst: any[] = []
-	if (response_type.terms && Array.isArray(response_type.terms)) {
-		for (const t of response_type.terms) {
-			const term: any = ds.cohort.termdb.q.termjsonByOneid(t)
-			if (!term) {
-				html += 'invalid term id:' + t + ' '
-			} else {
-				twLst.push({ id: term.id })
-			}
-		}
-	}
-
-	// Validate gene names — use geneExpression type for datasets with expression data,
-	// fall back to geneVariant for datasets with mutation data
-	if (response_type.geneNames && Array.isArray(response_type.geneNames)) {
-		for (const g of response_type.geneNames) {
-			const gene_hits = common_genes.filter(gene => gene == g.toLowerCase())
-			if (gene_hits.length == 0) {
-				html += 'invalid gene name:' + g + ' '
-			} else {
-				const geneName = g.toUpperCase()
-				if (dataset_json.hasGeneExpression) {
-					twLst.push({ term: { gene: geneName, type: 'geneExpression' } })
-				} else {
-					twLst.push({ term: { gene: geneName, name: geneName, type: 'geneVariant' } })
-				}
-			}
-		}
-	}
-
-	// Validate filters
-	if (response_type.simpleFilter && response_type.simpleFilter.length > 0) {
-		const validated_filters = validate_filter(response_type.simpleFilter, ds, '')
-		if (validated_filters.html.length > 0) {
-			html += validated_filters.html
-		} else {
-			pp_plot_json.filter = validated_filters.simplefilter
-		}
-	}
-
-	if (html.length > 0) {
-		return { type: 'html', html: html }
-	} else {
-		// Structure as termgroups matching what matrix.js expects:
-		// termgroups: [{ name: '', lst: [ { term: {...} }, ... ] }]
-		pp_plot_json.termgroups = [{ name: '', lst: twLst }]
-		return { type: 'plot', plot: pp_plot_json }
-	}
-}
-
-async function extract_samplescatter_terms_from_query(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	genes_list: string[],
-	ds: any,
-	testing: boolean
-) {
-	if (!dataset_json.prebuiltPlots || dataset_json.prebuiltPlots.length == 0) {
-		return { type: 'html', html: 'No pre-built scatter plots (t-SNE/UMAP) are available for this dataset' }
-	}
-
-	const Schema = {
-		$schema: 'http://json-schema.org/draft-07/schema#',
-		$ref: '#/definitions/SampleScatterType',
-		definitions: {
-			SampleScatterType: {
-				type: 'object',
-				properties: {
-					plotName: {
-						type: 'string',
-						description: 'Name of the pre-built scatter plot to display'
-					},
-					colorTW: {
-						type: ['string', 'null'],
-						description:
-							'Term name or gene name to overlay as color on the scatter plot. Set to null to remove the color overlay.'
-					},
-					shapeTW: {
-						type: ['string', 'null'],
-						description:
-							'Term name or gene name to overlay as shape on the scatter plot. Set to null to remove the shape overlay.'
-					},
-					term0: {
-						type: ['string', 'null'],
-						description:
-							'Term name to use for Z/Divide which splits the plot into panels. Set to null to remove the divide overlay.'
-					},
-					simpleFilter: {
-						type: 'array',
-						items: { $ref: '#/definitions/FilterTerm' },
-						description: 'Optional simple filter terms to restrict the sample set'
-					}
-				},
-				required: ['plotName'],
-				additionalProperties: false
-			},
-			...FILTER_TERM_DEFINITIONS
-		}
-	}
-
-	const common_genes = extractGenesFromPrompt(prompt, genes_list)
-
-	// Parse out training data from the dataset JSON
-	const scatter_ds = dataset_json.charts.find((chart: any) => chart.type == 'sampleScatter')
-	if (!scatter_ds) throw 'sampleScatter information is not present in the dataset file.'
-	if (scatter_ds.TrainingData.length == 0) throw 'No training data is provided for the sampleScatter agent.'
-
-	const training_data = formatTrainingExamples(scatter_ds.TrainingData)
-
-	const plotNames = dataset_json.prebuiltPlots.map((p: any) => p.name).join(', ')
-
-	let system_prompt =
-		'I am an assistant that extracts overlay parameters for pre-built scatter plots (t-SNE/UMAP). The final output must be in the following JSON format with NO extra comments. The JSON schema is as follows: ' +
-		JSON.stringify(Schema) +
-		' The available pre-built plots are: ' +
-		plotNames +
-		'. The "plotName" field must match one of these exactly. ' +
-		'The "colorTW", "shapeTW", and "term0" fields should contain names of clinical fields from the sqlite db OR gene names. ' +
-		'To remove an overlay, set the corresponding field to null explicitly. If the user does not mention a particular overlay, do NOT include that field in the output (omit it entirely). ' +
-		'Only include "colorTW", "shapeTW", or "term0" if the user explicitly mentions coloring, shaping, or dividing. ' +
-		FILTER_DESCRIPTION +
-		checkField(dataset_json.DatasetPrompt) +
-		checkField(scatter_ds.SystemPrompt) +
-		'\n The DB content is as follows: ' +
-		dataset_db_output.rag_docs.join(',') +
-		' training data is as follows:' +
-		training_data
-
-	if (dataset_json.hasGeneExpression && common_genes.length > 0) {
-		system_prompt += '\n List of relevant genes are as follows (separated by comma(,)):' + common_genes.join(',')
-	}
-
-	system_prompt += ' Question: {' + prompt + '} answer:'
-
-	const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-	if (testing) {
-		return { action: 'sampleScatter', response: JSON.parse(response) }
-	} else {
-		return validate_samplescatter_response(response, common_genes, dataset_json, ds)
-	}
-}
-
-function validate_samplescatter_response(response: string, common_genes: string[], dataset_json: any, ds: any) {
-	const response_type = JSON.parse(response)
-	let html = ''
-
-	if (response_type.html) html = response_type.html
-
-	// Validate plotName against prebuiltPlots
-	if (!response_type.plotName) {
-		html += 'plotName is required for sample scatter output'
-	} else {
-		const matchedPlot = dataset_json.prebuiltPlots.find(
-			(p: any) => p.name.toLowerCase() == response_type.plotName.toLowerCase()
-		)
-		if (!matchedPlot) {
-			const availablePlots = dataset_json.prebuiltPlots.map((p: any) => p.name).join(', ')
-			html += 'Unknown plot name: ' + response_type.plotName + '. Available plots are: ' + availablePlots
-		}
-	}
-
-	const pp_plot_json: any = {
-		chartType: 'sampleScatter',
-		name: response_type.plotName
-	}
-
-	// Helper to validate an overlay term (color, shape, or divide)
-	const validateOverlayTerm = (termName: string | null | undefined, fieldKey: string) => {
-		if (termName === null) {
-			// Explicit null means remove the overlay
-			pp_plot_json[fieldKey] = null
-			return
-		}
-		if (termName === undefined) {
-			// Not mentioned, don't include in output
-			return
-		}
-		const termValidation = validate_term(termName, common_genes, dataset_json, ds)
-		if (termValidation.html.length > 0) {
-			html += termValidation.html
-		} else {
-			const tw: any = { ...termValidation.term_type }
-			if (termValidation.category == 'float' || termValidation.category == 'integer') {
-				tw.q = { mode: 'continuous' }
-			}
-			pp_plot_json[fieldKey] = tw
-		}
-	}
-
-	validateOverlayTerm(response_type.colorTW, 'colorTW')
-	validateOverlayTerm(response_type.shapeTW, 'shapeTW')
-	validateOverlayTerm(response_type.term0, 'term0')
-
-	// Validate filters
-	if (response_type.simpleFilter && response_type.simpleFilter.length > 0) {
-		const validated_filters = validate_filter(response_type.simpleFilter, ds, '')
-		if (validated_filters.html.length > 0) {
-			html += validated_filters.html
-		} else {
-			pp_plot_json.filter = validated_filters.simplefilter
-		}
-	}
-
-	if (html.length > 0) {
-		return { type: 'html', html: html }
-	} else {
-		return { type: 'plot', plot: pp_plot_json }
-	}
-}
-
-function validate_term(response_term: string, common_genes: string[], dataset_json: any, ds: any) {
-	let html = ''
-	let term_type: any
-	let category: string = ''
-	const term: any = ds.cohort.termdb.q.termjsonByOneid(response_term)
-	if (!term) {
-		const gene_hits = common_genes.filter(gene => gene == response_term.toLowerCase())
-		if (gene_hits.length == 0) {
-			// Neither a clinical term nor a gene
-			html += 'invalid term id:' + response_term
-		} else {
-			if (dataset_json.hasGeneExpression) {
-				// Check to see if dataset support gene expression
-				term_type = { term: { gene: response_term.toUpperCase(), type: 'geneExpression' } }
-				category = 'float'
-			} else {
-				html += 'Dataset does not support gene expression'
-			}
-		}
-	} else {
-		term_type = { id: term.id }
-		category = term.type
-	}
-	return { term_type: term_type, html: html, category: category }
-}
-
-function countOccurrences(str: string, word: string): number {
-	if (word === '') return 0 // avoid infinite loops
-	let count = 0
-	let pos = 0
-
-	while ((pos = str.indexOf(word, pos)) !== -1) {
-		count++
-		pos += word.length // move past this match
-	}
-	return count
-}
-
-function removeLastOccurrence(str: string, word: string): string {
-	const index = str.lastIndexOf(word)
-	if (index === -1) return str // word not found
-
-	const occurrences = countOccurrences(str, word)
-	if (occurrences === 1) {
-		return str
-	} else {
-		// Slice out the word and concatenate the surrounding parts
-		return str.slice(0, index) + str.slice(index + word.length)
-	}
-}
-
-function sortSameCategoricalFilterKeys(filters: any[], ds: any): any {
-	let html = ''
-	const keys = filters.map(f => f.term)
-	if (new Set(keys).size == keys.length) return { filters: filters, html: html } // All filter terms have separate keys
-
-	const seen = new Set<string>()
-	const categorical_filter_terms_with_multiple_fields = new Set<string>()
-
-	for (const item of filters) {
-		if (seen.has(item.term)) categorical_filter_terms_with_multiple_fields.add(item.term)
-		else seen.add(item.term)
-	}
-
-	const multiple_fields_keys: { key: string; categories: string[] }[] = []
-	for (const key of categorical_filter_terms_with_multiple_fields) {
-		const term = ds.cohort.termdb.q.termjsonByOneid(key)
-		if (!term) {
-			html += 'invalid filter id:' + key
-		} else {
-			if (term.type == 'categorical') {
-				const multiple_fields = filters.filter(x => x.term == key)
-				multiple_fields_keys.push({ key: key, categories: multiple_fields.map(f => f.category) })
-			}
-		}
-	}
-
-	// Try to preserve the order of the original filter terms
-	const sorted_filter: FilterTerm[] = []
-	const seen2 = new Set<string>()
-	for (const f of filters) {
-		const repeated_term = multiple_fields_keys.find(x => x.key == f.term)
-		if (!repeated_term) {
-			sorted_filter.push(f)
-		} else {
-			if (!seen2.has(f.term)) {
-				const new_filter_term: { term: string; category: string[] } = {
-					term: f.term,
-					category: repeated_term.categories
-				}
-				seen2.add(f.term)
-				sorted_filter.push(new_filter_term)
-			}
-		}
-	}
-	return { filters: sorted_filter, html: html }
-}
-
-function validate_filter(filters: FilterTerm[], ds: any, group_name: string): any {
-	if (!Array.isArray(filters)) throw 'filter is not array'
-	const sorted_filters = sortSameCategoricalFilterKeys(filters, ds)
-	let filter_result: { simplefilter?: any; html: string } = { html: sorted_filters.html }
-	if (sorted_filters.filters.length <= 2) {
-		// If number of filter terms <=2 then simply a single iteration of generate_filter_term() is sufficient
-		const generated = generate_filter_term(sorted_filters.filters, ds)
-		filter_result.simplefilter = generated.simplefilter
-		filter_result.html += generated.html
-	} else {
-		if (sorted_filters.filters.length > num_filter_cutoff) {
-			filter_result.html +=
-				'For now, the maximum number of filter terms supported through the chatbot is ' + num_filter_cutoff
-			if (group_name.length > 0) {
-				// Group name is blank for summary filter, this is case for groups
-				filter_result.html +=
-					' . The number of filter terms for group ' + group_name + ' is ' + sorted_filters.filters.length + '\n' // Added temporary logic to restrict the number of filter terms to num_filter_cutoff.
-			} else {
-				// For summary filter prompts which do not have a group
-				filter_result.html += 'The number of filter terms for this query is ' + sorted_filters.filters.length
-			}
-		} else {
-			// When number of filter terms is greater than 2, then in each iteration the first two terms are taken and a filter object is created which is passed in the following iteration as a filter term
-			for (let i = 0; i < sorted_filters.filters.length - 1; i++) {
-				const filter_lst = [] as any[]
-				if (i == 0) {
-					filter_lst.push(sorted_filters.filters[i])
-				} else {
-					filter_lst.push(filter_result.simplefilter)
-				}
-				filter_lst.push(sorted_filters.filters[i + 1])
-				filter_result = generate_filter_term(filter_lst, ds)
-			}
-		}
-	}
-	return { simplefilter: filter_result.simplefilter, html: filter_result.html }
-}
-
-function generate_filter_term(filters: any, ds: any) {
-	let invalid_html = ''
-	const localfilter: any = { type: 'tvslst', in: true, lst: [] as any[] }
-	for (const f of filters) {
-		if (f.type == 'tvslst') {
-			localfilter.lst.push(f)
-		} else {
-			const term = ds.cohort.termdb.q.termjsonByOneid(f.term)
-			if (!term) {
-				invalid_html += 'invalid filter id:' + f.term
-			} else {
-				if (f.join) {
-					localfilter.join = f.join
-				}
-				if (term.type == 'categorical') {
-					if (Array.isArray(f.category)) {
-						// Array of categories
-						const categories: any[] = []
-						for (const category of f.category) {
-							const cat = findCategoryKey(term.values, category)
-							if (!cat) invalid_html += 'invalid category from ' + JSON.stringify(f)
-							else {
-								categories.push({ key: cat })
-							}
-						}
-						// term and category validated
-						localfilter.lst.push({
-							type: 'tvs',
-							tvs: {
-								term,
-								values: categories
-							}
-						})
-					} else {
-						// Single string category
-						const cat = findCategoryKey(term.values, f.category)
-						if (!cat) invalid_html += 'invalid category from ' + JSON.stringify(f)
-						else {
-							// term and category validated
-							localfilter.lst.push({
-								type: 'tvs',
-								tvs: {
-									term,
-									values: [{ key: cat }]
-								}
-							})
-						}
-					}
-				} else if (term.type == 'float' || term.type == 'integer') {
-					const numeric: any = {
-						type: 'tvs',
-						tvs: {
-							term,
-							ranges: []
-						}
-					}
-					const range: any = {}
-					if (f.start && !f.stop) {
-						range.start = Number(f.start)
-						range.stopunbounded = true
-					} else if (f.stop && !f.start) {
-						range.stop = Number(f.stop)
-						range.startunbounded = true
-					} else if (f.start && f.stop) {
-						range.start = Number(f.start)
-						range.stop = Number(f.stop)
-					} else {
-						invalid_html += 'Neither greater or lesser defined'
-					}
-					numeric.tvs.ranges.push(range)
-					localfilter.lst.push(numeric)
-				}
-			}
-		}
-	}
-	if (filters.length > 1 && !localfilter.join) {
-		localfilter.join = 'and' // Hardcoding and when the LLM is not able to detect the connection
-		//invalid_html += 'Connection (and/or) between the filter terms is not clear, please try to rephrase your question'
-	}
-	return { simplefilter: localfilter, html: invalid_html }
-}
-
-function findCategoryKey(termValues: Record<string, any>, category: string): string | undefined {
-	for (const ck in termValues) {
-		if (ck === category || termValues[ck].label === category) return ck
-	}
-	return undefined
-}
-
-async function parse_geneset_db(genedb: string) {
-	let genes_list: string[] = []
-	const db = new Database(genedb)
-	try {
-		// Query the database
-		const desc_rows = db.prepare('SELECT name from codingGenes').all()
-		desc_rows.forEach((row: any) => {
-			genes_list.push(row.name)
-		})
-		genes_list = genes_list.map(str => str.toLowerCase()) // Converting to lowercase
-	} catch (error) {
-		throw 'Could not parse geneDB' + error
-	} finally {
-		db.close()
-	}
-	return genes_list
-}
-
-async function parse_dataset_db(dataset_db: string) {
-	const db = new Database(dataset_db)
-	const rag_docs: string[] = []
-	const db_rows: DbRows[] = []
-	try {
-		// Query the database
-		const desc_rows = db.prepare('SELECT * from termhtmldef').all()
-
-		const description_map: any = []
-		// Process the retrieved rows
-		desc_rows.forEach((row: any) => {
-			const name: string = row.id
-			const jsonhtml = JSON.parse(row.jsonhtml)
-			const description: string = jsonhtml.description[0].value
-			description_map.push({ name: name, description: description })
-		})
-
-		const term_db_rows = db.prepare('SELECT * from terms').all()
-
-		term_db_rows.forEach((row: any) => {
-			const found = description_map.find((item: any) => item.name === row.id)
-			if (found) {
-				// Restrict db to only those items that have a description
-				const jsondata = JSON.parse(row.jsondata)
-				const description = description_map.filter((item: any) => item.name === row.id)
-				const term_type: string = row.type
-
-				const values: DbValue[] = []
-				if (jsondata.values && Object.keys(jsondata.values).length > 0) {
-					for (const key of Object.keys(jsondata.values)) {
-						const value = jsondata.values[key]
-						const db_val: DbValue = { key: key, value: value }
-						values.push(db_val)
-					}
-				}
-				const db_row: DbRows = {
-					name: row.id,
-					description: description[0].description,
-					values: values,
-					term_type: term_type
-				}
-				const stringified_db = parse_db_rows(db_row)
-				rag_docs.push(stringified_db)
-				db_rows.push(db_row)
-			}
-		})
-	} catch (error) {
-		throw 'Error in parsing dataset DB:' + error
-	} finally {
-		db.close()
-	}
-	return { db_rows: db_rows, rag_docs: rag_docs }
-}
-
-function parse_db_rows(db_row: DbRows) {
-	let output_string: string =
-		'Name of the field is:"' +
-		db_row.name +
-		'". This field is of the type:' +
-		db_row.term_type +
-		'. Description: ' +
-		db_row.description
-
-	if (db_row.values.length > 0) {
-		output_string += 'This field contains the following possible values.'
-		for (const value of db_row.values) {
-			if (value.value && value.value.label) {
-				output_string += 'The key is "' + value.key + '" and the label is "' + value.value.label + '".'
-			}
-		}
-	}
-	return output_string
 }
