--- server/routes/termdb.chat3.ts
@@ -1,285 +1,0 @@
-import type { ChatRequest, ChatResponse, LlmConfig, QueryClassification } from '#types'
-// import { ambiguousPoints } from '#types'
-import { mayLog } from '#src/helpers.ts'
-import { formatElapsedTime } from '#shared'
-import { readJSONFile, parse_geneset_db, getChatRelatedPlotTypes } from './chat/utils.ts'
-import { classifyQuery } from './chat/classify1.ts'
-import { classifyPlotType } from './chat/plot.ts'
-import { classifyNotPlot } from './chat/classify2.ts'
-import { inferScaffold } from './chat/scaffold.ts'
-import serverconfig from '../src/serverconfig.js'
-import { getDsAllowedTermTypes } from './termdb.config.ts'
-import { phrase2entity } from './chat/phrase2entity.ts'
-import { inferTermObjFromEntity } from './chat/entity2termObj.ts'
-import { resolveToTwTvs } from './chat/entity2twTvs.ts'
-import { answerDataQueries } from './chat/dataQueries.ts'
-import path from 'path'
-import fs from 'fs'
-import type { Scaffold, Phrase2EntityResult, SummaryScaffold } from './chat/scaffoldTypes.ts'
-import { resolveToPlotState } from './chat/scaffold2state.ts'
-
-export function init({ genomes }) {
-/*
-async function doOmnisearch(prompt: string, q, genome: any, ds: any) {
-	const matches = { equals: [], startsWith: [], startsWord: [], includes: [] }
-
-	// to allow search to work, must unescape special char, e.g. %20 to space
-	// const str = decodeURIComponent(q.findterm).toUpperCase()
-
-	let terms: any = []
-
-	let termdb:any
-	if (ds) {
-		// matches with dataset
-		if (ds?.cohort?.termdb) termdb = ds.cohort.termdb
-		else throw '.cohort.termdb not found on this dataset'
-	}
-	try {
-		console.log(`Running omnisearch for prompt: "${prompt}" on dataset: "${ds.label}"...`)
-		console.log(`q.cohortStr: "${q.cohortStr}", q.usecase: "${q.usecase}", q.treeFilter: ${JSON.stringify(q.treeFilter)}`)
-		const _terms = await termdb.q.findTermByName(prompt, q.cohortStr)
-		console.log(`Omnisearch found ${_terms.length} terms matching the prompt "${prompt}" in dataset "${ds.label}".`)
-		// console.log('Sample of matched terms:', _terms.slice(0, 5))
-		terms.push(..._terms.map(copy_term))
-	} catch (e) {
-		// if (e.stack) console.log(e.stack)
-		throw e 
-	}
-	return terms
-}
-*/
-	return async (req, res) => {
-		const q: ChatRequest = req.query
-		try {
-			const genome = genomes[q.genome]
-			if (!genome) throw 'invalid genome'
-			const ds = genome.datasets?.[q.dslabel]
-			if (!ds) throw 'invalid dslabel'
-			// check if ds supports termdb chat
-			if (!ds.queries.chat) {
-				return {
-					type: 'text',
-					text: 'Only search functionality supported for this data. No chat functionality supported.'
-				}
-			}
-
-			const aiFilesDir = serverconfig.binpath + '/../../dataset/ai/' + q.dslabel // This is the directory where the AI JSON files are stored for this dataset. This will use this as the base directory for resolving all agent file paths specified in the dataset JSON file.
-			let agentFiles: string[] = []
-			try {
-				// Read dataset JSON file
-				agentFiles = await fs.readdirSync(aiFilesDir).filter(file => file.endsWith('.json'))
-			} catch (err: any) {
-				if (err.code === 'ENOENT') throw new Error(`Directory not found: ${aiFilesDir}`)
-				if (err.code === 'ENOTDIR') throw new Error(`Path is not a directory: ${aiFilesDir}`)
-				throw err
-			}
-
-			const llm = serverconfig.llm
-			if (!llm) throw 'serverconfig.llm is not configured'
-			if (
-				llm.provider !== 'SJ' &&
-				llm.provider !== 'ollama' &&
-				llm.provider !== 'huggingface' &&
-				(llm.provider as string) !== 'azure'
-			) {
-				throw "llm.provider must be 'SJ', 'ollama', 'huggingface', or 'azure'"
-			}
-
-			// This toggles validation of LLM output. In this script, this will ALWAYS be false since we always want validation of LLM output,
-			// only for testing we set this variable to true
-			// const testing = false
-
-			// Access the cohort key/label
-			let rawFilter: any
-			if (typeof q.filter === 'string') {
-				try {
-					rawFilter = JSON.parse(q.filter)
-				} catch (e) {
-					throw new Error('Failed to parse filter JSON string: ' + e)
-				}
-			} else {
-				rawFilter = q.filter
-			}
-			const filter: any = rawFilter && typeof rawFilter === 'object' ? rawFilter : {}
-			const lst = Array.isArray(filter.lst) ? filter.lst : []
-			const cohortFilter = lst.find((item: any) => item.tag === 'cohortFilter')
-			const cohortKey = cohortFilter ? cohortFilter.tvs.values[0].key : ''
-			const supportedPlotTypes = ds.cohort.termdb.q?.getSupportedChartTypes(req)?.[cohortKey]
-			const chatSupportedPlotTypes = getChatRelatedPlotTypes(supportedPlotTypes)
-			const genedb = serverconfig.tpmasterdir + '/' + genome.genedb.dbfile
-			const allowedTermTypes = getDsAllowedTermTypes(ds) as string[]
-			const ai_output_json = await run_chat_pipeline(
-				q.prompt,
-				llm,
-				ds,
-				genedb,
-				agentFiles,
-				aiFilesDir,
-				chatSupportedPlotTypes,
-				allowedTermTypes,
-				genome
-				// 	testing
-			)
-			mayLog('From init: Final AI output JSON:', JSON.stringify(ai_output_json))
-			res.send(ai_output_json as ChatResponse)
-		} catch (e: any) {
-			if (e.stack) mayLog(e.stack)
-			res.send({ error: e?.message || e })
-		}
-	}
-}
-
-export async function run_chat_pipeline(
-	userPrompt: string,
-	llm: LlmConfig,
-	ds: any,
-	genedb: string,
-	agentFiles: string[],
-	aiFilesDir: string,
-	supportedPlotTypes: string[],
-	allowedTermTypes: string[],
-	genome: any
-	// testing: boolean
-) {
-	// Read main.json file
-	if (!fs.existsSync(path.join(aiFilesDir, 'main.json')))
-		throw 'Main data file is not specified for dataset:' + ds.label
-	const dataset_json: any = await readJSONFile(path.join(aiFilesDir, 'main.json'))
-
-	// Plot vs Not-plot classification
-	const time1 = new Date().valueOf()
-	const class_response: QueryClassification = await classifyQuery(userPrompt, llm)
-	mayLog('Time taken for classification:', formatElapsedTime(Date.now() - time1))
-
-	let ai_output_json: any
-	if (class_response.type === 'notplot') {
-		// If Not-plot: Resource or None classification
-		const time2 = new Date().valueOf()
-		const notPlotResult = await classifyNotPlot(userPrompt, llm, agentFiles, aiFilesDir) //, allowedTermTypes)
-		mayLog('Time taken for classify2:', formatElapsedTime(Date.now() - time2))
-
-		if (notPlotResult.type === 'html') {
-			ai_output_json = notPlotResult
-		} else {
-			ai_output_json = {
-				type: 'text',
-				text: 'Your query does not appear to be related to the available data visualizations. Please try rephrasing your question.'
-			}
-		}
-	} else if (class_response.type === 'binaryQuery') {
-		const answer = await answerDataQueries(userPrompt, llm, allowedTermTypes)
-		if (!answer) throw "Couldn't decide if this is data related query!"
-		mayLog('Data Binary Query: ', answer)
-		ai_output_json = answer
-	} else if (class_response.type === 'plot') {
-		let time = new Date().valueOf()
-		const plotType = await classifyPlotType(userPrompt, llm)
-		mayLog('Time taken to classify plot type:', formatElapsedTime(Date.now() - time))
-
-		// Check if the classified plot type is supported by this dataset
-		if (!supportedPlotTypes.includes(plotType)) {
-			const log = 'Plot type: "' + plotType + '" is not supported.'
-			ai_output_json = {
-				type: 'text',
-				text: log
-			}
-			mayLog(log)
-			return ai_output_json
-		}
-
-		const genes_list = await parse_geneset_db(genedb)
-
-		// If supported plot type, figure out the scaffold according to the plot type
-		mayLog('#################################################')
-		mayLog('####### First phase: Infer Plot Scaffolds #######')
-		mayLog('#################################################')
-		time = new Date().valueOf()
-		const dataset_db = serverconfig.tpmasterdir + '/' + ds.cohort.db.file
-		const scaffoldResult = await inferScaffold(
-			userPrompt,
-			plotType,
-			llm,
-			genome,
-			genes_list,
-			allowedTermTypes,
-			dataset_json,
-			ds,
-			dataset_db
-		)
-		mayLog('ScaffoldResult: ', scaffoldResult)
-		if (
-			(plotType === 'hiercluster' && 'plot' in scaffoldResult && scaffoldResult.type === 'plot') ||
-			('text' in scaffoldResult && scaffoldResult.type === 'text')
-		) {
-			// In case of geneExpression clustering, the plot state is generated through a single LLM call and invoking downstream steps is not necessary.
-			return scaffoldResult
-		}
-		mayLog('Time taken to infer scaffold:', formatElapsedTime(Date.now() - time))
-
-		if (!scaffoldResult)
-			throw 'Scaffold result is empty or undefined, which is unexpected. Please check the inferScaffold agent for potential issues.'
-		if ('type' in scaffoldResult && scaffoldResult.type === 'text') {
-			return scaffoldResult // Return msg/error
-		}
-
-		const subplotType =
-			(scaffoldResult as Scaffold).plotType === 'summary' ? (scaffoldResult as SummaryScaffold).chartType : undefined
-
-		mayLog('#################################################')
-		mayLog("####### Second phase: From Scaffolds's phrases infer Entities #######")
-		mayLog('#################################################')
-		time = new Date().valueOf()
-		const phrase2entityResult = await phrase2entity(
-			scaffoldResult as Scaffold,
-			plotType,
-			llm,
-			genes_list,
-			dataset_json,
-			ds,
-			genome
-		)
-		mayLog('Time taken to phrase 2 entity:', formatElapsedTime(Date.now() - time))
-		if ('type' in phrase2entityResult && phrase2entityResult.type === 'text') {
-			return phrase2entityResult // Return msg/error
-		}
-		mayLog(phrase2entityResult)
-
-		mayLog('#################################################')
-		mayLog('####### Third phase: From Entities infer Term Objects #######')
-		mayLog('#################################################')
-		time = new Date().valueOf()
-		const termObj = await inferTermObjFromEntity(
-			phrase2entityResult as Phrase2EntityResult,
-			plotType,
-			llm,
-			dataset_db,
-			genes_list,
-			genome
-		)
-		mayLog('Time taken to infer term objects:', formatElapsedTime(Date.now() - time))
-		mayLog('Inferred termObj from entity:', JSON.stringify(termObj))
-
-		mayLog('#################################################')
-		mayLog('####### Fourth phase: From Term Objects to TwTvs Objects #######')
-		mayLog('#################################################')
-		time = new Date().valueOf()
-		const twTvsObj = await resolveToTwTvs(termObj, plotType, llm, dataset_db, genome)
-		mayLog('Time taken to resolve to TwTvs object from termObj:', formatElapsedTime(Date.now() - time))
-		if ('type' in twTvsObj && twTvsObj.type === 'text') {
-			return twTvsObj // Return msg/error
-		}
-		mayLog('twTvsObj:', twTvsObj)
-
-		mayLog('#################################################')
-		mayLog('####### Fifth/Final phase: From TwTvs Objects to Plot States #######')
-		mayLog('#################################################')
-		time = new Date().valueOf()
-		ai_output_json = resolveToPlotState(twTvsObj, plotType, subplotType)
-		mayLog('Time taken to resolve to plot state:', formatElapsedTime(Date.now() - time))
-		// TODO: might need a validation step here to check if the scaffoldResult contains valid term types that
-		// are present in the dataset and compatible with the plot type, and if not return an error message to the user.
-		// This is a bit complex because it requires cross-referencing the inferred scaffold with the dataset's schema and allowed term types,
-		//  but it would improve robustness.
-	}
-	return ai_output_json
-}
